refactor: replace Ember Data attempt with naïve route implemtnation
Authoring the adapter and serializer needed to get Ember Data working correctly is going to take more time than I want to use. Reverting to using a simpler method of interaction for now.

An alternative solution could have been to fix the API to send a true JSON API format response, I may look into this in the future.